requirements.txt
cycler==0.10.0
graphviz==0.17
kiwisolver==1.3.1
matplotlib==3.4.2
neat-python==0.92
numpy==1.21.0
Pillow==8.3.1
pygame==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
